+++ inst/doc/ssgraph-example.R
@@ -15,10 +15,14 @@ data.sim <- bdgraph.sim( n = 100, p = 8, graph = "scale-free", vis = TRUE )
 round( head( data.sim $ data, 4 ), 2 )
 
 ## ----fig.align = 'center'-----------------------------------------------------
-ssgraph.obj <- ssgraph( data = data.sim, method = "ggm", iter = 5000, save = TRUE )
+ssgraph.obj <- ssgraph( data = data.sim, method = "ggm", iter = 5000, 
+                        save = TRUE, verbose = FALSE )
 
 summary( ssgraph.obj )
 
 ## ----fig.align = 'center'-----------------------------------------------------
 compare( data.sim, ssgraph.obj, main = c( "Target", "ssgraph" ), vis = TRUE )
+
+## ----fig.align = 'center'-----------------------------------------------------
+plotroc( ssgraph.obj, data.sim, cut = 200 )